requirements.txt
# Requirements for AI-Based Child Safety System

# Core dependencies
pyserial==3.5
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pywhatkit==5.4
PyAutoGUI==0.9.54
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Configuration management
python-dotenv==1.0.0

# Additional dependencies for face_recognition
dlib==19.24.2
cmake==3.27.7

# Note: Installation order matters for some packages
# Install in this order:
# 1. pip install cmake
# 2. pip install dlib
# 3. pip install face-recognition
# 4. pip install -r requirements.txt (for the rest)
